# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Indie List API",
    "version": "1.0.0",
    "description": "Documentação oficial da API do Indie List"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "GetRatingsResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "stars": {
                  "type": "number"
                },
                "description": {
                  "type": "string",
                  "nullable": true
                },
                "userName": {
                  "type": "string"
                },
                "contentTitle": {
                  "type": "string"
                },
                "totalRatings": {
                  "type": "number"
                },
                "updatedAt": {
                  "type": "string"
                },
                "createdAt": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                },
                "data": {
                  "type": "null",
                  "example": null
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                },
                "data": {
                  "type": "null",
                  "example": null
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/rating": {
      "get": {
        "operationId": "get-rating",
        "summary": "",
        "description": "Endpoint para buscar avaliações de conteúdos do usuário, podendo filtrar por limite de resultados",
        "tags": [
          "Rating"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna uma lista de avaliações de conteúdos do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetRatingsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  }
}